requirements.txt
requests>=2.31.0
Pillow>=10.0.0
pillow-avif-plugin>=1.4.0
imageio-ffmpeg>=0.5.1
edge-tts>=6.1.12
pyttsx3>=2.90
deep-translator>=1.11.4
duckduckgo-search>=6.2.0
ddgs>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
playwright>=1.48.0
